Refactor: validation.js을 utils 파일에서 validation 파일로 옮김
